% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num-to-words.R
\name{decimals_to_words}
\alias{decimals_to_words}
\title{Convert a sequence of digits to a sequence of words}
\usage{
decimals_to_words(digits)
}
\arguments{
\item{digits}{single character with just digits e.g. "1056"}
}
\value{
single character string e.g. "point one zero five six"
}
\description{
This function trims traling zeros and prefixes the result with the word "point"
}
