% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num-to-words.R
\name{num_to_words}
\alias{num_to_words}
\title{Convert numbers to words}
\usage{
num_to_words(num, decimals = 3)
}
\arguments{
\item{num}{numeric vector}

\item{decimals}{number of decimal places}
}
\value{
character vector
}
\description{
Convert numbers to words
}
