% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num-to-words.R
\name{small_int_to_words}
\alias{small_int_to_words}
\title{convert a number in range [0,999] to words}
\usage{
small_int_to_words(num)
}
\arguments{
\item{num}{integer in range 0-999}
}
\description{
convert a number in range [0,999] to words
}
