% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num-to-words.R
\name{triplets_to_words}
\alias{triplets_to_words}
\title{convert a character vector of triplets into a word description}
\usage{
triplets_to_words(triplets)
}
\arguments{
\item{triplets}{characcter vector of number triplets}
}
\description{
convert a character vector of triplets into a word description
}
\examples{
\dontrun{
triplets_to_words(c('123', '345'))
}
}
