% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words-to-num.R
\name{words_to_num}
\alias{words_to_num}
\title{Convert a character strings into numeric values}
\usage{
words_to_num(words)
}
\arguments{
\item{words}{character vector}
}
\value{
numeric vector
}
\description{
Convert a character strings into numeric values
}
\examples{
\dontrun{
words_to_num(c("twelve", "one hundred point three"))
}

}
