% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words-to-num.R
\name{words_to_num_single}
\alias{words_to_num_single}
\title{Convert a single character string into a numeric value}
\usage{
words_to_num_single(words)
}
\arguments{
\item{words}{single character string}
}
\value{
numeric value
}
\description{
Convert a single character string into a numeric value
}
\examples{
\dontrun{
words_to_num_single("twelve")
}

}
